+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [0.9.0]
+## [0.9.1]
 
 ### Fixed
 
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - `LongPollingServiceContext` now generic.
 - Use channel instead of callbacks.
 - Make `SendError` copy.
+- Intervals and timeouts accept Duration.
 
 ## [0.8.0]